Ransomware detection model - production version 2
