+++ examples/ransomware_detection/README.md
@@ -43,7 +43,7 @@ docker run --rm -ti --gpus=all -p8000:8000 -p8001:8001 -p8002:8002 -v $PWD/model
    tritonserver --model-repository=/models/triton-model-repo \
                 --exit-on-error=false \
                 --model-control-mode=explicit \
-                --load-model ransomw-model-short-rf
+                --load-model ransomware_model_tl
 ```
 
 ##### Verify Model Deployment
@@ -53,7 +53,7 @@ Once Triton server finishes starting up, it will display the status of all loade
 +----------------------------+---------+--------+
 | Model                      | Version | Status |
 +----------------------------+---------+--------+
-| ransomw-model-short-rf     | 1       | READY  |
+| ransomware_model_tl     | 1       | READY  |
 +----------------------------+---------+--------+
 ```
 
@@ -72,10 +72,12 @@ Run the following from the `examples/ransomware_detection` directory to start th
 ```bash
 python run.py --server_url=localhost:8001 \
               --sliding_window=3 \
-              --model_name=ransomw-model-short-rf \
+              --model_name=ransomware_model_tl \
               --conf_file=./config/ransomware_detection.yaml \
-              --input_glob=${MORPHEUS_ROOT}/examples/data/appshield/*/snapshot-*/*.json \
-              --output_file=./ransomware_detection_output.jsonlines
+              --input_topic=ransomware_input \
+              --output_topic=ransomware_output \
+              --bootstrap_servers broker:9092 \
+              --group_id ransomware_group
 ```
 
 Input features for a short model can be taken from every three snapshots sequence, such as (1, 2, 3), or (2, 3, 4). The sliding window represents the number of subsequent snapshots that need to be taken into consideration when generating the input for a model. Sliding window for the medium model is `5` and for the long model it is `10`.
@@ -108,20 +110,13 @@ Options:
   --server_url TEXT               Tritonserver url  [required]
   --sliding_window INTEGER RANGE  Sliding window to be used for model input
                                   request  [x>=1]
-  --input_glob TEXT               Input glob pattern to match files to read.
+  --input_topic TEXT              Input Kafka topic for receiving the 
+                                  data [required]
+  --output_topic TEXT             Output Kafka topic for receiving the 
+                                  data [required]
+  --bootstrap_servers TEXT        Kafka bootstrap server [required]
                                   For example,
-                                  './input_dir/*/snapshot-*/*.json' would read
-                                  all files with the 'json' extension in the
-                                  directory 'input_dir'.  [required]
-  --watch_directory BOOLEAN       The watch directory option instructs this
-                                  stage to not close down once all files have
-                                  been read. Instead it will read all files
-                                  that match the 'input_glob' pattern, and
-                                  then continue to watch the directory for
-                                  additional files. Any new files that are
-                                  added that match the glob will then be
-                                  processed.
-  --output_file TEXT              The path to the file where the inference
-                                  output will be saved.
+                                  broker:9092
+  --group_id TEXT                 Kafka group_id topic [required]
   --help                          Show this message and exit.
   ```
