--- examples/ransomware_detection/models/ransomw-model-short-rf/config.pbtxt
@@ -1,52 +1,0 @@
-name: "ransomw-model-short-rf"
-backend: "fil"
-max_batch_size: 8192
-input [
- {
-    name: "input__0"
-    data_type: TYPE_FP32
-    dims: [ 297 ]
-  }
-]
-output [
- {
-    name: "output__0"
-    data_type: TYPE_FP32
-    dims: [ 2 ]
-  }
-]
-instance_group [{ kind: KIND_GPU }]
-parameters [
-  {
-    key: "model_type"
-    value: { string_value: "treelite_checkpoint" }
-  },
-  {
-    key: "predict_proba"
-    value: { string_value: "true" }
-  },
-  {
-    key: "output_class"
-    value: { string_value: "true" }
-  },
-  {
-    key: "threshold"
-    value: { string_value: "0.5" }
-  },
-  {
-    key: "algo"
-    value: { string_value: "ALGO_AUTO" }
-  },
-  {
-    key: "storage_type"
-    value: { string_value: "AUTO" }
-  },
-  {
-    key: "blocks_per_sm"
-    value: { string_value: "0" }
-  }
-]
-
-dynamic_batching {
-  max_queue_delay_microseconds: 100
-}
