+++ examples/ransomware_detection/models/ransomware_model_tl/config.pbtxt
