+++ examples/ransomware_detection/stages/appshield_partitioner.py
